Added logic to calculate resource production. Each time a base is retrieved it calculates the new resources amount based on the last time the resources were updated.

+++ base-manager/src/main/java/org/example/config/WorldConfig.java
@@ -18,5 +18,5 @@ public class WorldConfig {
 
     private Double EXPONENTIAL;
 
-    private Integer DEFAULT_AMOUNT;
+    private Double DEFAULT_AMOUNT;
 }
